# ArzTiServer/Controllers/RezeptIDApi.cs
/*
 * Webservice ArzTI API
 *
 * Kommunikation des Webservers mit den jeweiligen Rechenzentren * UC: Abruf der eRezept-Daten - neue (noch nicht gesendete) | eRezept-ID | Zeitraum * UC: Abruf der eMuster16-Daten - neue (noch nicht gesendete) | eRezept-ID | Zeitraum * UC: Abruf der pRezept-Daten - neue (noch nicht gesendete) | eRezept-ID | Zeitraum * UC: Aktualisierung des eRezept-Status - eRezept-ID * UC: Löschen eines eRezepts - eRezept-ID * UC: Aktualisierung von Kundendaten (Apotheke -> Rechenzentrum Zuordnung) - Apotheke-IK * UC: ... 
 *
 * OpenAPI spec version: V1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using ArzTiServer.Attributes;
using ArzTiServer.Security;
using ArzTiServer.Models;

namespace IO.Swagger.Controllers
{
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class RezeptIDApiController : ControllerBase
    {
        /// <summary>
        /// löscht ein Rezept
        /// </summary>
        /// <remarks>deleteRezeptId</remarks>
        /// <param name="apoik">Die Apotheken IK.</param>
        /// <param name="reztyp">.</param>
        /// <param name="rezid">Die RezeptID.</param>
        /// <response code="200">Liefert das Rezept für die ID</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="404">The specified resource was not found</response>
        [HttpDelete]
        [Route("/v1/apotheke/{apoik}/rezept/{reztyp}/{rezid}")]
        [Authorize(AuthenticationSchemes = BasicAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("DeleteRezeptId")]
        [SwaggerResponse(statusCode: 200, type: typeof(Rezept), description: "Liefert das Rezept für die ID")]
        [SwaggerResponse(statusCode: 401, type: typeof(Error), description: "Unauthorized")]
        [SwaggerResponse(statusCode: 404, type: typeof(Error), description: "The specified resource was not found")]
        public virtual IActionResult DeleteRezeptId([FromRoute][Required] string apoik, [FromRoute][Required] RezeptTyp reztyp, [FromRoute][Required] string rezid)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Rezept));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(Error));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(Error));
            string exampleJson = null;
            exampleJson = "{\n  \"uid\" : {\n    \"ruid\" : \"ruid\"\n  },\n  \"data\" : \"data\",\n  \"apoik\" : \"apoik\",\n  \"id\" : {\n    \"typ\" : \"ERezept\",\n    \"id\" : \"id\"\n  },\n  \"status\" : \"status\"\n}";

            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Rezept>(exampleJson)
            : default(Rezept);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// liefert ein Rezept
        /// </summary>
        /// <remarks>getRezeptId</remarks>
        /// <param name="apoik">Die Apotheken IK.</param>
        /// <param name="reztyp">.</param>
        /// <param name="rezid">Die RezeptID.</param>
        /// <response code="200">Liefert das Rezept für die ID</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="404">The specified resource was not found</response>
        [HttpGet]
        [Route("/v1/apotheke/{apoik}/rezept/{reztyp}/{rezid}")]
        [Authorize(AuthenticationSchemes = BasicAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("GetRezeptId")]
        [SwaggerResponse(statusCode: 200, type: typeof(Rezept), description: "Liefert das Rezept für die ID")]
        [SwaggerResponse(statusCode: 401, type: typeof(Error), description: "Unauthorized")]
        [SwaggerResponse(statusCode: 404, type: typeof(Error), description: "The specified resource was not found")]
        public virtual IActionResult GetRezeptId([FromRoute][Required] string apoik, [FromRoute][Required] RezeptTyp reztyp, [FromRoute][Required] string rezid)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Rezept));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(Error));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(Error));
            string exampleJson = null;
            exampleJson = "{\n  \"uid\" : {\n    \"ruid\" : \"ruid\"\n  },\n  \"data\" : \"data\",\n  \"apoik\" : \"apoik\",\n  \"id\" : {\n    \"typ\" : \"ERezept\",\n    \"id\" : \"id\"\n  },\n  \"status\" : \"status\"\n}";

            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Rezept>(exampleJson)
            : default(Rezept);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Statusabfrage eines Rezeptes
        /// </summary>
        /// <remarks>getRezeptIdStatus</remarks>
        /// <param name="apoik">Die Apotheken IK.</param>
        /// <param name="reztyp">.</param>
        /// <param name="rezid">Die Rezept-ID.</param>
        /// <response code="200">Liefert den Status des Rezeptes für die ID</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="404">The specified resource was not found</response>
        [HttpGet]
        [Route("/v1/apotheke/{apoik}/rezept/{reztyp}/{rezid}/status")]
        [Authorize(AuthenticationSchemes = BasicAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("GetRezeptIdStatus")]
        [SwaggerResponse(statusCode: 200, type: typeof(Rezept), description: "Liefert den Status des Rezeptes für die ID")]
        [SwaggerResponse(statusCode: 401, type: typeof(Error), description: "Unauthorized")]
        [SwaggerResponse(statusCode: 404, type: typeof(Error), description: "The specified resource was not found")]
        public virtual IActionResult GetRezeptIdStatus([FromRoute][Required] string apoik, [FromRoute][Required] RezeptTyp reztyp, [FromRoute][Required] string rezid)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Rezept));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(Error));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(Error));
            string exampleJson = null;
            exampleJson = "{\n  \"uid\" : {\n    \"ruid\" : \"ruid\"\n  },\n  \"data\" : \"data\",\n  \"apoik\" : \"apoik\",\n  \"id\" : {\n    \"typ\" : \"ERezept\",\n    \"id\" : \"id\"\n  },\n  \"status\" : \"status\"\n}";

            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Rezept>(exampleJson)
            : default(Rezept);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Statusänderung eines Rezeptes
        /// </summary>
        /// <remarks>patchRezeptIdStatus</remarks>
        /// <param name="apoik">Die Apotheken IK.</param>
        /// <param name="reztyp">.</param>
        /// <param name="rezid">Die Rezept-ID.</param>
        /// <param name="status">Status value</param>
        /// <response code="200">Liefert den Status des Rezeptes für die ID</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="404">The specified resource was not found</response>
        [HttpPatch]
        [Route("/v1/apotheke/{apoik}/rezept/{reztyp}/{rezid}")]
        [Authorize(AuthenticationSchemes = BasicAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("PatchRezeptIdStatus")]
        [SwaggerResponse(statusCode: 200, type: typeof(RezeptStatus), description: "Liefert den Status des Rezeptes für die ID")]
        [SwaggerResponse(statusCode: 401, type: typeof(Error), description: "Unauthorized")]
        [SwaggerResponse(statusCode: 404, type: typeof(Error), description: "The specified resource was not found")]
        public virtual IActionResult PatchRezeptIdStatus([FromRoute][Required] string apoik, [FromRoute][Required] RezeptTyp reztyp, [FromRoute][Required] string rezid, [FromQuery][Required()] string status)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(RezeptStatus));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(Error));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(Error));
            string exampleJson = null;
            exampleJson = "{\n  \"uid\" : {\n    \"ruid\" : \"ruid\"\n  },\n  \"id\" : {\n    \"typ\" : \"ERezept\",\n    \"id\" : \"id\"\n  },\n  \"status\" : \"status\"\n}";

            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<RezeptStatus>(exampleJson)
            : default(RezeptStatus);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}