# ArzTiServer/Models/RezeptTyp.cs
/*
 * Webservice ArzTI API
 *
 * Kommunikation des Webservers mit den jeweiligen Rechenzentren * UC: Abruf der eRezept-Daten - neue (noch nicht gesendete) | eRezept-ID | Zeitraum * UC: Abruf der eMuster16-Daten - neue (noch nicht gesendete) | eRezept-ID | Zeitraum * UC: Abruf der pRezept-Daten - neue (noch nicht gesendete) | eRezept-ID | Zeitraum * UC: Aktualisierung des eRezept-Status - eRezept-ID * UC: Löschen eines eRezepts - eRezept-ID * UC: Aktualisierung von Kundendaten (Apotheke -> Rechenzentrum Zuordnung) - Apotheke-IK * UC: ... 
 *
 * The version of the OpenAPI document: V1
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using ArzTiServer.Converters;

namespace ArzTiServer.Models
{ 
        /// <summary>
        /// E-, M-, oder P-Rezept
        /// </summary>
        /// <value>E-, M-, oder P-Rezept</value>
        [TypeConverter(typeof(CustomEnumConverter<RezeptTyp>))]
        [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
        public enum RezeptTyp
        {
            
            /// <summary>
            /// Enum ERezeptEnum for ERezept
            /// </summary>
            [EnumMember(Value = "ERezept")]
            ERezeptEnum = 1,
            
            /// <summary>
            /// Enum MRezeptEnum for MRezept
            /// </summary>
            [EnumMember(Value = "MRezept")]
            MRezeptEnum = 2,
            
            /// <summary>
            /// Enum PRezeptEnum for PRezept
            /// </summary>
            [EnumMember(Value = "PRezept")]
            PRezeptEnum = 3
        }
}