# ArzTiServer/wwwroot/swagger-original.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Webservice ArzTI API",
    "description": "Kommunikation des Webservers mit den jeweiligen Rechenzentren ",
    "version": "V1"
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Local server"
    },
    {
      "url": "http://arzsoftware.de/v1",
      "description": "Main (production) server"
    },
    {
      "url": "http://staging-api.arzsoftware.de",
      "description": "Internal staging server for testing"
    }
  ],
  "paths": {
...
    "/apotheke/{apoik}/rezept/{reztyp}/{rezid}": {
      "get": {
        "tags": [
          "RezeptID"
        ],
        "summary": "liefert ein Rezept",
        "description": "",
        "operationId": "getRezeptId",
        "parameters": [
          {
            "name": "apoik",
            "in": "path",
            "required": true,
            "description": "Die Apotheken IK.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reztyp",
            "in": "path",
            "required": true,
            "description": ".",
            "schema": {
              "$ref": "#/components/schemas/RezeptTyp"
            }
          },
          {
            "name": "rezid",
            "in": "path",
            "required": true,
            "description": "Die RezeptID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert das Rezept für die ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rezept"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Rezept"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "RezeptID"
        ],
        "summary": "löscht ein Rezept",
        "description": "",
        "operationId": "deleteRezeptId",
        "parameters": [
          {
            "name": "apoik",
            "in": "path",
            "required": true,
            "description": "Die Apotheken IK.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reztyp",
            "in": "path",
            "required": true,
            "description": ".",
            "schema": {
              "$ref": "#/components/schemas/RezeptTyp"
            }
          },
          {
            "name": "rezid",
            "in": "path",
            "required": true,
            "description": "Die RezeptID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert das Rezept für die ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rezept"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Rezept"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_write"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "RezeptID"
        ],
        "summary": "Statusänderung eines Rezeptes",
        "description": "Statusänderung eines Rezeptes",
        "operationId": "patchRezeptIdStatus",
        "parameters": [
          {
            "name": "apoik",
            "in": "path",
            "required": true,
            "description": "Die Apotheken IK.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reztyp",
            "in": "path",
            "required": true,
            "description": ".",
            "schema": {
              "$ref": "#/components/schemas/RezeptTyp"
            }
          },
          {
            "name": "rezid",
            "in": "path",
            "required": true,
            "description": "Die Rezept-ID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "description": "Status value",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert den Status des Rezeptes für die ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RezeptStatus"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/RezeptStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      }
    },
    "/apotheke/{apoik}/rezept/{reztyp}/{rezid}/status": {
      "get": {
        "tags": [
          "RezeptID"
        ],
        "summary": "Statusabfrage eines Rezeptes",
        "description": "Statusabfrage eines Rezeptes",
        "operationId": "getRezeptIdStatus",
        "parameters": [
          {
            "name": "apoik",
            "in": "path",
            "required": true,
            "description": "Die Apotheken IK.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reztyp",
            "in": "path",
            "required": true,
            "description": ".",
            "schema": {
              "$ref": "#/components/schemas/RezeptTyp"
            }
          },
          {
            "name": "rezid",
            "in": "path",
            "required": true,
            "description": "Die Rezept-ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert den Status des Rezeptes für die ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rezept"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Rezept"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "RezeptStatus": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RezeptId"
          },
          "uid": {
            "$ref": "#/components/schemas/RezeptUId"
          },
          "status": {
            "type": "string",
            "description": "Status des Rezepts"
          }
        }
      },
      "RezeptTyp": {
        "type": "string",
        "enum": [
          "ERezept",
          "MRezept",
          "PRezept"
        ]
      },
      "Rezept": {
        "title": "Rezept",
        "type": "object",
        "description": "base64 codierte Leistung",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RezeptId"
          },
          "uid": {
            "$ref": "#/components/schemas/RezeptUId"
          },
          "data": {
            "type": "string"
          }
        }
      },
      "RezeptId": {
        "title": "Rezept",
        "type": "object",
        "description": "die Rezept ID",
        "properties": {
          "typ": {
            "$ref": "#/components/schemas/RezeptTyp"
          },
          "id": {
            "type": "string"
          }
        }
      },
      "RezeptUId": {
        "title": "Rezept",
        "type": "object",
        "description": "die Rezept RUID",
        "properties": {
          "ruid": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
...
}